[Issue-3556] Split file
