+++ packages/extension-base/src/types/account/action/derive.ts
@@ -1,6 +1,8 @@
 // Copyright 2019-2022 @subwallet/extension-base authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+import { KeypairType } from '@subwallet/keyring/types';
+
 export interface CreateDeriveAccountInfo {
   name: string;
   suri: string;
@@ -53,4 +55,25 @@ export interface RequestGetDeriveSuggestion {
 export interface ResponseGetDeriveSuggestion {
   proxyId: string;
   suri: string;
+}
+
+export interface DeriveInfo {
+  suri?: string;
+  deriveIndex?: number;
+}
+
+export interface NextDerivePair {
+  deriveIndex: number;
+  suri: string;
+}
+
+export interface IDerivePathInfo_ {
+  raw: string;
+  type: KeypairType;
+  suri: string;
+  depth: number;
+}
+
+export interface DerivePathInfo extends Omit<IDerivePathInfo_, 'type'> {
+  type: KeypairType | 'unified';
 }
