+++ packages/extension-base/src/utils/account/derive/index.ts
@@ -1,0 +1,5 @@
+// Copyright 2019-2022 @subwallet/extension-base authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+export * from './info';
+export * from './validate';
